Add reference detail tabs
