+++ client/src/js/nav/components/Sidebar.js
@@ -43,7 +43,7 @@ const Sidebar = (props) => {
         case "References":
             return (
                 <Nav className="sidebar">
-                    <SidebarItem title="Manage" link="/refs" icon="th-list" />
+                    <SidebarItem title="Browse" link="/refs" icon="th-list" />
                     <SidebarItem title="Indexes" link="/refs/indexes" icon="tags" />
                     <SidebarItem title="Settings" link="/refs/settings" icon="cogs" />
                 </Nav>
