+++ client/src/js/references/components/Detail/Index.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ReferenceIndex = () => (
+    <div>Index</div>
+);
+
+export default ReferenceIndex;
