+++ client/src/js/references/components/Detail/OTU.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ReferenceOTU = () => (
+    <div>OTU</div>
+);
+
+export default ReferenceOTU;
